Put ATTNWebViewHandlerUITests into a different file and do not put it in the unit test scheme
